Split resolver enum function payload tests
