+++ src/typechecker/tests/resolver_struct_enum_metadata/enum_metadata.rs
@@ -1,5 +1,7 @@
 use super::*;
 
+mod function_payloads;
+
 #[test]
 fn check_program_with_symbols_validates_resolver_enum_variant_payload_counts() {
     let program = parse_program(
@@ -76,89 +78,6 @@ Option: Some(i32), None
             .message
             .contains("resolver variant symbol 'Some' has payload type 'bool', expected 'i32'")),
         "expected resolver enum variant payload type diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_with_symbols_validates_resolver_enum_function_type_payloads() {
-    let program = parse_program(
-        r#"
-Callback: Wrap((i32) i32), None
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.set_variant_payload_type_name_for_test(
-        Namespace::Variant,
-        "Wrap",
-        Some("i32".to_string()),
-    );
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program_with_symbols(&program, &symbols)
-        .expect_err("resolver enum function type payload mismatch should fail");
-
-    assert!(
-        err.iter().any(|d| d.message.contains(
-            "resolver variant symbol 'Wrap' has payload type 'i32', expected '(i32) i32'"
-        )),
-        "expected resolver enum function type payload diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_with_symbols_validates_resolver_enum_typed_payload_metadata() {
-    let program = parse_program(
-        r#"
-Callback: Wrap((i32) i32), None
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.set_variant_payload_type_for_test(Namespace::Variant, "Wrap", Some(AstType::I32));
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program_with_symbols(&program, &symbols)
-        .expect_err("resolver typed enum payload metadata mismatch should fail");
-
-    assert!(
-        err.iter().any(|d| d.message.contains(
-            "resolver variant symbol 'Wrap' has typed payload type 'i32', expected '(i32) i32'"
-        )),
-        "expected resolver typed enum payload diagnostic, got {err:?}"
-    );
-}
-
-#[test]
-fn check_program_with_symbols_validates_resolver_generic_enum_function_type_payloads() {
-    let program = parse_program(
-        r#"
-Callback<T>: Wrap((T) T), None
-"#,
-    );
-    let mut symbols = crate::resolver::Resolver::new()
-        .resolve_program(&program)
-        .expect("resolver succeeds");
-    symbols.set_variant_payload_type_name_for_test(
-        Namespace::Variant,
-        "Wrap",
-        Some("T".to_string()),
-    );
-    let mut tc = TypeChecker::new();
-
-    let err = tc
-        .check_program_with_symbols(&program, &symbols)
-        .expect_err("resolver generic enum function type payload mismatch should fail");
-
-    assert!(
-        err.iter().any(|d| d
-            .message
-            .contains("resolver variant symbol 'Wrap' has payload type 'T', expected '(T) T'")),
-        "expected resolver generic enum function type payload diagnostic, got {err:?}"
     );
 }
